refactor: Migrate metadata to __version__
