+++ akcli/__version__.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+
+"""
+Package metadata.
+"""
+
+__title__ = "akcli"
+__url__ = ""
+__description__ = "Command-line interface to make requests to Akamai API."
+__epilog__ = f"See documentation at: [u]{__url__}[/u] for more information."
+__version__ = "v0.1.0"
+__author__ = "Javi C. Guerrero"
+__author_email__ = "jcorreag@pm.me"
+__license__ = "MIT"
